Add building example
Creating scriptable objects on the fly and patching a few things, it is
possible to add new buildings at runtime as illustrated in this mod. It
has some issues ranging from rendering & serialization to simulation
precision. It is not a good illustration of a good building but rather a
proof of concept that new ones can be added